LuckyCardMinigame&DailyReward&QuizScreen&ConnectionError&ImageBlinkBagFix

+++ FF Diamond 1.0/Assets/Scripts/UI/ViewSystem/IUIViewController.cs
@@ -3,7 +3,7 @@ namespace UI.ViewSystem
     public interface IUIViewController
     {
         void ShowScreen(UIScreenId id);
-        void ShowPopup(UIPopupId id, int reward, string result = "");
+        void ShowPopup(UIPopupId id, float reward, string result = "");
         void HidePopup(UIPopupId id);
         void HideAllPopups();
     }
